backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.3.223
opencv-python==4.8.1.78
numpy>=1.24.4,<2.0.0
Pillow==10.1.0
pydantic>=2.9.2
# DeepSORT dependencies
scipy>=1.9.0
filterpy>=1.4.5

